weather_man/requirements.txt
requests
openai
langchain
langchain_openai
langchain_core
langgraph.prebuilt
python-dotenv
langchain_community
faiss-cpu